[Security Solution] Relocate final sourcerer model exports and remove folder
Move the last three exports out of `sourcerer/store/model.ts` and delete the
now-empty `sourcerer/` folder:

- `initDataView` moves to `data_view_manager/utils/init_data_view.ts` with an
  inline object type, dropping the `SourcererDataView` interface (its only
  remaining consumer).
- `SourcererUrlState` is renamed to `DataViewManagerUrlState` and moved to
  `data_view_manager/redux/types.ts`.
- `SelectedDataView` moves to `use_ti_data_view.ts`, its sole consumer.
