+++ x-pack/solutions/security/plugins/security_solution/public/data_view_manager/mocks/timeline_data_view.ts
@@ -7,7 +7,7 @@
 
 import type { DataViewSpec } from '@kbn/data-views-plugin/public';
 import type { BrowserFields } from '@kbn/timelines-plugin/common';
-import { initDataView } from '../../sourcerer/store/model';
+import { initDataView } from '../utils/init_data_view';
 import { mockBrowserFields, mockIndexFields } from '../../common/containers/source/mock';
 import {
   DEFAULT_DATA_VIEW_ID,
