+++ x-pack/solutions/security/plugins/security_solution/public/data_view_manager/redux/types.ts
@@ -6,6 +6,17 @@
  */
 
 import type { DataViewSpec } from '@kbn/data-views-plugin/common';
+import type { PageScope } from '../constants';
+
+/**
+ * Shape of the data view manager selection persisted in the URL, keyed by page scope.
+ */
+export type DataViewManagerUrlState = Partial<{
+  [id in PageScope]: {
+    id: string;
+    selectedPatterns: string[];
+  };
+}>;
 
 export interface ScopedDataViewSelectionState {
   dataViewId: string | null;
