+++ x-pack/solutions/security/plugins/security_solution/public/threat_intelligence/modules/indicators/hooks/use_ti_data_view.ts
@@ -6,13 +6,37 @@
  */
 
 import { useMemo } from 'react';
-import type { SelectedDataView } from '../../../../sourcerer/store/model';
+import type { DataViewSpec } from '@kbn/data-views-plugin/public';
+import type { BrowserFields as DataViewBrowserFields } from '@kbn/timelines-plugin/common';
 import { useSelectedPatterns } from '../../../../data_view_manager/hooks/use_selected_patterns';
 import { useBrowserFields } from '../../../../data_view_manager/hooks/use_browser_fields';
 import { RawIndicatorFieldId } from '../../../../../common/threat_intelligence/types/indicator';
 import { DESCRIPTION } from './translations';
 import { useDataView } from '../../../../data_view_manager/hooks/use_data_view';
 import type { BrowserFields } from '../../../types';
+
+/**
+ * Combined data from a Kibana data view and its selected patterns to create
+ * selected data view state.
+ */
+export interface SelectedDataView {
+  /**
+   * @deprecated use EcsFlat or fields / indexFields from data view
+   */
+  browserFields: DataViewBrowserFields;
+  dataViewId: string | null; // null if legacy pre-8.0 timeline
+  /** do the selected indices exist */
+  indicesExist: boolean;
+  /** is an update being made to the data view */
+  loading: boolean;
+  /** all selected patterns */
+  selectedPatterns: string[];
+  /**
+   * Easier to add this additional data rather than
+   * try to extend the SelectedDataView type from DataView.
+   */
+  sourcererDataView: DataViewSpec;
+}
 
 /**
  * Inline definition for a runtime field "threat.indicator.name" we are adding for indicators grid
